(maint) Split option parser into it's own file
